1. Find Password
Detective Buckshee Junior has been approached by the shanthiniketan kids society for help in finding the password to the games complex. After hearing the scenario, detective Buckshee Junior realizes that he will need a programmer’s support. He contacts you and requests your help. Please help the detective by writing a function to generate the password.
The Scenario is as below
•	Five numbers are available with the kids.
•	These numbers are either stable or unstable.
•	A number is stable if each of its digit occur the same number of times, i.e. the frequency of each digit in the number is the same. For e.g. 2277,4004,11,23,583835,1010 are examples of stable numbers.
•	Similarly, a number is unstable if the frequency of each digit in the number is NOT the same.
For e.g 221,4314,101,233,58135 are examples of unstable numbers.

The password can be found as below:
•	i.e. password =sum of all unstable numbers

Assuming that the five numbers are passed to a function as input1, input2, input3, input4 and input5, complete the function to find and return the password.

•	For Example: If input1=12, imput2=1313, input3=122, input4=678 and input5=898.
•	Stable numbers are 12, 1313 and 678
•	Unstable numbers are 122 and 898

So, the password should be sum of all unstable numbers = 1020

public int findPassword(int input1, int input2, int input3, int input4, int input5) {
       --------------------------------------------------------------------------------------------------
}