Duplicate Count
Guru is having set of integers and he want to know the first repeating element and its occurrence in it. He is felt, it’s difficult to find manually. So, write a code in java and help him to find first repeating element and its occurrence. An array of integers is given as input and you need to find and print the first repeating element and its occurrence.
Note:
1.    If the given input array is not having the repeating element, then print NODUPLICATEELEMENT
2.    Array contains only positive elements
Input Format:
    First line contains an integer N(size of the array) and second line contains N elements separated by space(array elements).
Output Format:
    A line contains first repeating element and its occurrence which are separated by a space.